+++ manpage.go
@@ -1,0 +1,8 @@
+package main
+
+import (
+	_ "embed"
+)
+
+//go:embed autoreverse.8
+var Manpage []byte